refactor: server handling from pompt
